Frame duration is an integer

+++ demo/src/fire.py
@@ -103,7 +103,7 @@ class Fire:
         if not self.okay:
             return
 
-        self.led_strip.set_frame_duration(1000.0 / config.FIRE_FRAME_RATE)
+        self.led_strip.set_frame_duration(int(1000.0 / config.FIRE_FRAME_RATE))
 
     def frame_rendered(self, _):
         self.frame_upload()
